Fix history showing max time for first random move

+++ src/Board.java
@@ -203,7 +203,7 @@ public class Board {
         return opponent;
     }
 
-    public double getCurPlayerTurnTimeLeft() {
+    public int getCurPlayerTurnTimeLeft() {
         return curPlayerTurnTimeLeft;
     }
